- Move api_client implementation to a separate file - Add methods for configuring an APIClient
